TerminalSettings: Split TerminalSettingsWidget into different files
